Bootstrap two basic Plug apps (Alice and Bob)
In order to use two separate web applications we prepared a couple of
basic apps – named Alice and Bob.

They are extremely basic and their only purpose is to start a
Plug.Cowboy adapter on port 4000 and respond to the request to the index
URL with a line of text (distinguishable from each other).