Introduce Router app (using Plug.Router)
With these changes we introduced our main component – a parent
application that works as a reverse proxy (or as a router) to our basic
Alice and Bob mini-apps.

It is based on Plug.Router and forwards the requests to the
corresponding modules.

One important change that we had to introduce to mini-apps is about
adding a configuration option that helps us to control if Alice and Bob
apps are going to start their own Cowboy servers: by default they run
it, but when included and compiled within the Router application – we
disable webservers. Basically, because we want to run Router.Dispatcher
plug on the same port.